core: Machines can provide a custom local.conf
Modify setup-script-internal to check if a machine offers a local.conf file and
if so append its contents to auto.conf. As this file is rewritten on each call to
the script, it should be kept up to date with the configuration changes.
This allows the machines to provide additional default configurations which can
easily be changed by the user if needed.

Change-Id: I9d435fc158babc3ede4dbb387b5a7c23df50cc93
Signed-off-by: Teo Couprie Diaz <teo.coupriediaz@arm.com>